deps: migrate antlr-rust 0.3 -> antlr4rust 0.5.2; automate parser codegen
Swap the parser runtime from the unmaintained `antlr-rust 0.3.0-beta` to
`antlr4rust 0.5.2`. Behavior-preserving: same vendored v2.5 grammar, submodule
pointer unchanged.

Regenerate the lexer/parser/visitor/listener with the antlr4rust tool and
migrate the compiler to the 0.5.2 API:
- token-type constants now live on the lexer module (not the parser);
- antlr4rust emits no per-channel constant, so the COMMENTS channel is resolved
  by name from `channelNames`;
- token indices / intervals are `i32` rather than `isize`;
- `get_dfa_string` is now required on `TokenSource` (delegated by IndentAwareLexer);
- `ParseTreeWalker::walk` now returns `Result` (the walk is infallible for these
  listeners, which surface user errors via `CompilerError`, so it is `expect`ed).

Replace hand-patching of the generated files with a `generate_parser` codegen
binary (`cargo run -p yarnspinner_codegen --bin generate_parser`): it runs the
ANTLR tool and applies the required fix-ups -- the `ParserParserContext` rename,
the `visit_node` infinite-recursion disambiguation, the `new_with_text`
injection, and deletion of the unused base listener/visitor -- each guarded by an
occurrence-count assertion that fails loudly if a future antlr4rust release
changes the generated output. The old `type` -> `declaration_type` patch is gone:
the generator emits a valid `type_` field. The generated/readme.md is rewritten
to document the single command instead of a manual checklist.

Notes on the two upstream workarounds:
- `antlr_rust_ext.rs` is kept and documented. Neither antlr4rust nor dbt-antlr4
  exposes the hidden-channel `TokenStream` methods it reimplements
  (`get_hidden_tokens_to_left/right`), so it cannot be removed by the dep swap;
  upstreaming those is a follow-up.
- `hashable_interval` still derives the source interval by hand: antlr4rust 0.5
  dropped the `get_source_interval` override antlr-rust 0.3 had, so the default
  returns the invalid `(-1, -2)` interval for every parser rule context.

Other:
- The raw generated-lexer token stream for `significant_whitespace.yarn` gains
  one trailing `TEXT TEXT NEWLINE` group under antlr4rust 0.5.2 (closer to the C#
  reference); the golden test is updated.
- Pin the generated files and Yarn test fixtures to `eol=lf` in `.gitattributes`
  so codegen and golden-output tests are deterministic regardless of `autocrlf`.

+++ crates/compiler/src/visitors/code_generation_visitor.rs
@@ -5,9 +5,9 @@ use crate::prelude::generated::yarnspinnerlexer;
 use crate::prelude::generated::yarnspinnerparser::*;
 use crate::prelude::generated::yarnspinnerparservisitor::YarnSpinnerParserVisitorCompat;
 use crate::prelude::*;
-use antlr_rust::parser_rule_context::ParserRuleContext;
-use antlr_rust::token::Token;
-use antlr_rust::tree::{ParseTree, ParseTreeVisitorCompat, Tree};
+use antlr4rust::parser_rule_context::ParserRuleContext;
+use antlr4rust::token::Token;
+use antlr4rust::tree::{ParseTree, ParseTreeVisitorCompat, Tree};
 use std::ops::Deref;
 use std::rc::Rc;
 use yarnspinner_core::prelude::OpCode;
@@ -31,7 +31,7 @@ impl<'a, 'input: 'a> CodeGenerationVisitor<'a, 'input> {
             _dummy: Default::default(),
         }
     }
-    pub(crate) fn token_to_operator(token: isize) -> Option<Operator> {
+    pub(crate) fn token_to_operator(token: i32) -> Option<Operator> {
         // operators for the standard expressions
         match token {
             yarnspinnerlexer::OPERATOR_LOGICAL_LESS_THAN_EQUALS => {
